Move shared JavaScript FFI helpers into prelude
